Minor cleanup of SAI's VectorTypeTest
* Remove accidentally committed test
  that is not vector specfic and is
  already covered elsewhere.
* Relocate deletion tests to the
  VectorUpdateDeleteTest.
